requirements.txt
maya==0.6.1
requests==2.18.4
python-dotenv==0.12.0